Moving structs into types package
