+++ types/autogen_subdomain.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 type AutogenSubdomain struct {
 	IdModel
