+++ types/layer.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 type Layer string
