+++ types/module.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 type Module struct {
 	UidCreatedModel
@@ -6,7 +6,7 @@ type Module struct {
 	Name        string       `json:"name"`
 	Description string       `json:"description"`
 	IsPublic    bool         `json:"isPublic"`
-	Layer       Layer        `json:"layer"`
+	Layer       Layer    `json:"layer"`
 	Category    CategoryName `json:"category"`
 	Type        string       `json:"type"`
