+++ types/module_version.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 import (
 	"github.com/google/uuid"
