+++ types/outputs.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 type Outputs map[string]OutputItem
