+++ types/stack.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 type Stack struct {
 	IdModel
