+++ types/subdomain.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 type Subdomain struct {
 	IdModel
