+++ types/uid_created_model.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 import (
 	"github.com/google/uuid"
