+++ types/workspace_target.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 import (
 	"fmt"
